Repo: Clean up includes, outsource file load/save
