+++ include/ui.h
@@ -2,7 +2,6 @@
 #define UI_H
 
 #include <session.h>
-#include <screen.h>
 
 bool fe_safe_file_dialog( Session *s );
 bool fe_quit_dialog( Session *s );
